price-control/requirements.txt
pytibber==0.24.0
paho-mqtt==1.6.1
